Splices SearchView, removes 'Toast' for now
